fix(dialog): make "Install" button default
Ensures that if a user uses the CLI or file manager to open a flatpakref
file to install, they can press Enter to actually install the file.

Note that this will not work if a user initiates the installation
via the main Repoman window because they need to click the
file chooser button in order to choose which file to install, and this
action sets focus on the file chooser button. GTK always sends the
activate signal to the focused widget, and only sends to the default
widget if there is no widget with focus.